requirements.txt
sentence-transformers
numpy
chromadb
python-dotenv
openai
supabase